pyproject.toml
[tool.poetry]
name = "foreal"
version = "2022.6.1"
description = "Data processing for real data"
authors = ["Matthias Meyer <contact@matthiasmeyer.xyz>"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.4.1"
appdirs = "^1.4.4"
dask = "^2021.3.1"
zarr = {git = "https://github.com/shikharsg/zarr", rev = "chunk_cache"}
xarray = {git = "https://github.com/niowniow/xarray"}
scipy = "1.6"
obspy = "^1.2.2"
tqdm = "^4.60.0"
frozendict = "2.0.6"
python-benedict = "^0.24.0"
distributed = "^2021.6.2"
pandas = "^1.3.2"
dash-mantine-components = "^0.9.0"
dash = "^2.5.0"
dash-bootstrap-components = "^1.1.0"
dash-cytoscape = "^0.3.0"
holoviews = "^1.14.9"
compress-pickle = "^2.1.0"
numpy = "1.22"

[tool.poetry.dev-dependencies]
pre-commit = "^2.13.0"
pytest = "^6.2.4"
mkdocs = "^1.3.0"
pydoc-markdown = {extras = ["novella"], version = "^4.6.3"}
novella = "^0.2.3"
mkdocs-material = "^8.3.6"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
